everything in action

+++ ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: mwilk <mwilk@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/11/10 14:22:41 by mwilk             #+#    #+#             */
-/*   Updated: 2014/11/10 15:08:49 by mwilk            ###   ########.fr       */
+/*   Updated: 2016/05/09 16:40:24 by mwilk            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 char	*ft_strjoin(const char *s1, const char *s2)
 {
-	char	*str;
-	size_t	size;
+	char			*str;
+	size_t			size;
 
 	if (s1 == NULL || s2 == NULL)
 		return (NULL);
-	size = ft_strlen(s1) + ft_strlen(s2);
+	size = (size_t)ft_strlen(s1) + (size_t)ft_strlen(s2);
 	str = (char *)malloc(sizeof(char) * (size + 1));
 	if (str == NULL)
 		return (NULL);
