+++ tt_malloc_intab.c
@@ -6,7 +6,7 @@
 /*   By: mwilk <mwilk@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/04/01 23:00:36 by mwilk             #+#    #+#             */
-/*   Updated: 2016/04/01 23:18:11 by mwilk            ###   ########.fr       */
+/*   Updated: 2016/05/09 16:18:45 by mwilk            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,10 +17,10 @@ int		**tt_malloc_intab(int j, int i)
 	int		**tabint;
 	int		bzero;
 
-	tabint = (int **)malloc(sizeof(int *) * j);
+	tabint = (int **)malloc(sizeof(int *) * (size_t)j);
 	while (--j != -1)
 	{
-		tabint[j] = (int *)malloc(sizeof(int) * i);
+		tabint[j] = (int *)malloc(sizeof(int) * (size_t)i);
 		bzero = -1;
 		while (++bzero < i)
 			tabint[j][bzero] = 0;
