+++ tt_pnbl.c
@@ -6,7 +6,7 @@
 /*   By: mwilk <mwilk@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/11/13 14:26:26 by mwilk             #+#    #+#             */
-/*   Updated: 2015/02/04 19:30:32 by mwilk            ###   ########.fr       */
+/*   Updated: 2016/05/09 16:21:58 by mwilk            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void    pnbl(const char *s)
 {
-    int print;
-    
-    print = ft_strlen(s);
-    pnbr(print);
-    pel("");
+	int	print;
+
+	print = (int)ft_strlen(s);
+	pnbr(print);
+	pel("");
 }
